feat: Add profile picture and bio to user registration
This commit adds support for inserting a profile picture and bio when registering a new user. The `INSERT INTO Admins` statement has been updated to include the `ProfilePicture` and `Bio` columns. The profile picture is set to a default image ('default.jpg') and the bio is set to a greeting message using the username.