feat: Add DEFAULT_BOT_PASSWORD environment variable
This commit adds the `DEFAULT_BOT_PASSWORD` environment variable to the project. The `DEFAULT_BOT_PASSWORD` is used to set the default password for the bot user, which is responsible for controlling the application and performing administrative tasks. The variable can be set either through the command line or by using a `.env` file in the project's root directory.